hidos-django/hidos/requirements/base.txt
celery>=3.1.23
docutils>=0.12
urllib3>=1.16
psycopg2>=2.6.2
numpy==1.11.1
matplotlib>=1.5.1
Markdown>=2.6.6
Pillow>=3.3.0
Django[bcrypt]>=1.10
django-extensions>=1.7.1
django-rest-auth[with_social]>=0.8.1
django-model-utils>=2.5
django-vanilla-views>=1.0.4
django-filter>=0.13.0
django-material>=0.8.0
django-crispy-forms>=1.6.0
djangorestframework>=3.4.3
djangorestframework-jwt>=1.8.0

scikit-image>=0.12.3
scipy==0.17.1
mahotas>=1.4.1
pymorph>=0.96
